+++ packages/instrumentation-typeorm/src/utils/suppressTracing.ts
@@ -1,0 +1,14 @@
+import { createContextKey, Context } from '@opentelemetry/api';
+
+const SUPPRESS_TYPEORM_INTERNAL_TRACING_KEY = createContextKey(
+    'instrumentation-typeorm Context Key SUPPRESS_TYPEORM_INTERNAL_TRACING'
+);
+
+export const suppressTypeormInternalTracing = (context: Context) =>
+    context.setValue(SUPPRESS_TYPEORM_INTERNAL_TRACING_KEY, true);
+
+export const isTypeormInternalTracingSuppressed = (context: Context) =>
+    context.getValue(SUPPRESS_TYPEORM_INTERNAL_TRACING_KEY) === true;
+
+export const unsuppressTypeormInternalTracing = (context: Context) =>
+    context.deleteValue(SUPPRESS_TYPEORM_INTERNAL_TRACING_KEY);
