+++ packages/instrumentation-typeorm/test/Repository.spec.ts
@@ -1,33 +1,11 @@
 import 'mocha';
 import expect from 'expect';
-import { TypeormInstrumentation, TypeormInstrumentationConfig } from '../src';
+import { TypeormInstrumentation } from '../src';
 import { getTestSpans } from 'opentelemetry-instrumentation-testing-utils';
+
 const instrumentation = new TypeormInstrumentation();
-
+import { localPostgreSQLOptions, User } from './utils';
 import * as typeorm from 'typeorm';
-
-@typeorm.Entity()
-export class User {
-    @typeorm.PrimaryGeneratedColumn()
-    id: number;
-
-    @typeorm.Column()
-    firstName: string;
-
-    @typeorm.Column()
-    lastName: string;
-
-    @typeorm.Column({ default: true })
-    isActive: boolean;
-}
-const options: typeorm.ConnectionOptions = {
-    type: 'postgres',
-    host: 'localhost',
-    port: 5432,
-    username: 'motti',
-    password: 'mysecretpassword',
-    entities: [User],
-};
 
 describe('Repository', () => {
     beforeEach(() => {
@@ -38,8 +16,8 @@ describe('Repository', () => {
         instrumentation.disable();
     });
 
-    it('findAndCount enableInternalInstrumentation:false', async () => {
-        const conn = await typeorm.createConnection(options);
+    it('findAndCount', async () => {
+        const conn = await typeorm.createConnection(localPostgreSQLOptions);
         try {
             const repo = conn.getRepository(User);
             const [users, count] = await repo.findAndCount();
@@ -47,22 +25,6 @@ describe('Repository', () => {
 
             const spans = getTestSpans();
             expect(spans.length).toEqual(1);
-        } finally {
-            await conn.close();
-        }
-    });
-
-    it('findAndCount enableInternalInstrumentation:true', async () => {
-        const config: TypeormInstrumentationConfig = { enableInternalInstrumentation: true };
-        instrumentation.setConfig(config);
-        const conn = await typeorm.createConnection(options);
-        try {
-            const repo = conn.getRepository(User);
-            const [users, count] = await repo.findAndCount();
-            expect(count).toBeGreaterThan(0);
-
-            const spans = getTestSpans();
-            expect(spans.length).toEqual(2);
         } finally {
             await conn.close();
         }
